Feature/add playlist and fix peter links

+++ deploy.sh
@@ -1,25 +1,34 @@
 #!/bin/bash
-echo 'Please have hugo installed: https://gohugo.io/'
-#cp docs/README.md hugo/content/index.md
+# deploys to cbmministry site as hugo simply generates static files, we still need to copy to webserver
+# use "./deploy.sh y" to skip confirmation
+
+hugo_exists=$(which hugo)
+if [[ -z $hugo_exists ]]; then
+    echo 'Please have hugo installed: https://gohugo.io/'
+    exit 1
+fi
+
+# for scripts to skip confirmation
+sync_immediately=$1
 
 # Clean the public folder before rebuild - not sure why needed but issues on some machine
 rm -rf  hugo/public/*
 
-# Generating the hugo site
+echo "Generating hugo site"
 cd hugo
 hugo
 
 # Uploading generated site to 
 cd ../
 # open hugo/public/index.html # On mac, launch browser to check
-echo "You can use the following to upload::"
-echo "    rsync -avz hugo/public/ a1:www/cbmministry/wed_study/ "
-echo "---------------------------------------------------------"
 
-# Confirm whether to perform sync now
-read -p "Woulc you like to run the above rsync now (y/n)? " -n 1 -r
-echo    # (optional) move to a new line
-if [[ $REPLY =~ ^[Yy]$ ]]
+if [[ $sync_immediately != "y" ]]; then
+    # Confirm whether to perform sync now
+    read -p "Would you like to run the above rsync now (y/n)? " -n 1 -r
+    echo    # (optional) move to a new line
+fi
+
+if [[ $REPLY =~ ^[Yy]$ || $sync_immediately == "y" ]]
 then
     rsync -avz hugo/public/ aonemak1@a1make.net:www/cbmministry/wed_study/
 fi
